pyproject.toml
[project]
name = "api-chat-and-message"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=6.0.1",
    "djangorestframework>=3.16.1",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.1",
]

[dependency-groups]
dev = [
    "django-debug-toolbar>=6.1.0",
    "pre-commit>=4.5.1",
]
